+++ src/write_delim.cpp
@@ -4,22 +4,22 @@ using namespace Rcpp;
 
 // Defined later to make copyright clearer
 template <class Stream>
-void stream_delim(Stream& output, Rcpp::RObject x, int i, char delim);
+void stream_delim(Stream& output, Rcpp::RObject x, int i, char delim, const std::string& na);
 
 template <class Stream>
-void stream_delim_row(Stream& output, Rcpp::List x, int i, char delim) {
+void stream_delim_row(Stream& output, Rcpp::List x, int i, char delim, const std::string& na) {
   int p = Rf_length(x);
 
   for (int j = 0; j < p; ++j) {
-    stream_delim(output, x[j], i, delim);
+    stream_delim(output, x[j], i, delim, na);
     if (j != p - 1)
       output << delim;
   }
   output << '\n';
 }
 
-bool needs_quote(const char* string, char delim) {
-  if (*string == 'N' && *(string + 1) == 'A')
+bool needs_quote(const char* string, char delim, const std::string& na) {
+  if (string == na)
     return true;
 
   for (const char* cur = string; *cur != '\0'; ++cur) {
@@ -31,8 +31,8 @@ bool needs_quote(const char* string, char delim) {
 }
 
 template <class Stream>
-void stream_delim(Stream& output, const char* string, char delim) {
-  bool quotes = needs_quote(string, delim);
+void stream_delim(Stream& output, const char* string, char delim, const std::string& na) {
+  bool quotes = needs_quote(string, delim, na);
 
   if (quotes)
     output << '"';
@@ -52,7 +52,8 @@ void stream_delim(Stream& output, const char* string, char delim) {
 }
 
 template <class Stream>
-void stream_delim(Stream& output, List df, char delim, bool col_names = true, bool append = false) {
+void stream_delim(Stream& output, List df, char delim, const std::string& na,
+                  bool col_names = true, bool append = false) {
   int p = Rf_length(df);
   if (p == 0)
     return;
@@ -60,7 +61,7 @@ void stream_delim(Stream& output, List df, char delim, bool col_names = true, bo
   if (col_names) {
     CharacterVector names = as<CharacterVector>(df.attr("names"));
     for (int j = 0; j < p; ++j) {
-      stream_delim(output, names, j, delim);
+      stream_delim(output, names, j, delim, na);
       if (j != p - 1)
         output << delim;
     }
@@ -71,16 +72,17 @@ void stream_delim(Stream& output, List df, char delim, bool col_names = true, bo
   int n = Rf_length(first_col);
 
   for (int i = 0; i < n; ++i) {
-    stream_delim_row(output, df, i, delim);
+    stream_delim_row(output, df, i, delim, na);
   }
 }
 
 // [[Rcpp::export]]
-std::string stream_delim(List df, std::string path, char delim, bool col_names = true, bool append = false) {
+std::string stream_delim(List df, const std::string& path, char delim, const std::string& na,
+                         bool col_names = true, bool append = false) {
   if (path == "") {
     std::ostringstream output;
     output.precision(17);
-    stream_delim(output, df, delim, col_names, append);
+    stream_delim(output, df, delim, na, col_names, append);
     return output.str();
   } else {
     std::ofstream output(path.c_str(), append ? std::ofstream::app : std::ofstream::trunc);
@@ -88,7 +90,7 @@ std::string stream_delim(List df, std::string path, char delim, bool col_names =
       stop("Failed to open '%s'.", path);
     }
     output.precision(17);
-    stream_delim(output, df, delim, col_names, append);
+    stream_delim(output, df, delim, na, col_names, append);
     return "";
   }
 }
@@ -99,7 +101,7 @@ std::string stream_delim(List df, std::string path, char delim, bool col_names =
 // License: GPL-2
 
 template <class Stream>
-void stream_delim(Stream& output, RObject x, int i, char delim) {
+void stream_delim(Stream& output, RObject x, int i, char delim, const std::string& na) {
   switch (TYPEOF(x)) {
   case LGLSXP: {
     int value = LOGICAL(x)[i];
@@ -108,14 +110,14 @@ void stream_delim(Stream& output, RObject x, int i, char delim) {
     } else if (value == FALSE) {
       output << "FALSE";
     } else {
-      output << "NA";
+      output << na;
     }
     break;
   }
   case INTSXP: {
     int value = INTEGER(x)[i];
     if (value == NA_INTEGER) {
-      output << "NA";
+      output << na;
     } else {
       output << value;
     }
@@ -125,7 +127,7 @@ void stream_delim(Stream& output, RObject x, int i, char delim) {
     double value = REAL(x)[i];
     if (!R_FINITE(value)) {
       if (ISNA(value)) {
-        output << "NA";
+        output << na;
       } else if (ISNAN(value)) {
         output << "NaN";
       } else if (value > 0) {
@@ -141,9 +143,9 @@ void stream_delim(Stream& output, RObject x, int i, char delim) {
   case STRSXP: {
     RObject value = STRING_ELT(x, i);
     if (value == NA_STRING) {
-      output << "NA";
+      output << na;
     } else {
-      stream_delim(output, Rf_translateCharUTF8(STRING_ELT(x, i)), delim);
+      stream_delim(output, Rf_translateCharUTF8(STRING_ELT(x, i)), delim, na);
     }
     break;
   }